renamed into calendarPageJS
